+++ vscode/src/edit/execute.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode'
 import type { ChatEventSource, ContextFile, ContextMessage } from '@sourcegraph/cody-shared'
 
 import type { EditIntent, EditMode } from './types'
+import type { EditSupportedModels } from './prompt'
 import type { FixupTask } from '../non-stop/FixupTask'
 
 export interface ExecuteEditArguments {
@@ -13,6 +14,7 @@ export interface ExecuteEditArguments {
     intent?: EditIntent
     range?: vscode.Range
     mode?: EditMode
+    model?: EditSupportedModels
 }
 
 /**