+++ vscode/src/non-stop/FixupController.ts
@@ -1,16 +1,14 @@
 import * as vscode from 'vscode'
 
 import {
-    displayPath,
     displayPathBasename,
     type ChatEventSource,
     type ContextFile,
     type ContextMessage,
 } from '@sourcegraph/cody-shared'
 
-import type { ExecuteEditArguments } from '../edit/execute'
+import { executeEdit } from '../edit/execute'
 import type { EditIntent, EditMode } from '../edit/types'
-import { getEditSmartSelection } from '../edit/utils/edit-selection'
 import { logDebug } from '../log'
 import { telemetryService } from '../services/telemetry'
 import { telemetryRecorder } from '../services/telemetry-v2'
@@ -26,23 +24,14 @@ import { NewFixupFileMap, type FixupFile } from './FixupFile'
 import { FixupFileObserver } from './FixupFileObserver'
 import { FixupScheduler } from './FixupScheduler'
 import { FixupTask, type taskID } from './FixupTask'
-import { FixupTypingUI } from './FixupTypingUI'
-import type {
-    FixupFileCollection,
-    FixupIdleTaskRunner,
-    FixupTaskFactory,
-    FixupTextChanged,
-} from './roles'
+import type { FixupFileCollection, FixupIdleTaskRunner, FixupTextChanged } from './roles'
 import { CodyTaskState } from './utils'
+import type { EditSupportedModels } from '../edit/prompt'
+import { getInput } from '../edit/input/get-input'
 
 // This class acts as the factory for Fixup Tasks and handles communication between the Tree View and editor
 export class FixupController
-    implements
-        FixupFileCollection,
-        FixupIdleTaskRunner,
-        FixupTaskFactory,
-        FixupTextChanged,
-        vscode.Disposable
+    implements FixupFileCollection, FixupIdleTaskRunner, FixupTextChanged, vscode.Disposable
 {
     private tasks = new Map<taskID, FixupTask>()
     private readonly files: FixupFileObserver
@@ -52,7 +41,6 @@ export class FixupController
     private readonly decorator = new FixupDecorator()
     private readonly codelenses = new FixupCodeLenses(this)
     private readonly contentStore = new ContentProvider()
-    private readonly typingUI = new FixupTypingUI(this)
 
     private _disposables: vscode.Disposable[] = []
 
@@ -171,10 +159,44 @@ export class FixupController
     }
 
     public async promptUserForTask(
-        args: ExecuteEditArguments,
+        document: vscode.TextDocument,
+        range: vscode.Range,
+        expandedRange: vscode.Range | undefined,
+        mode: EditMode,
+        model: EditSupportedModels,
+        intent: EditIntent,
+        contextMessages: ContextMessage[],
         source: ChatEventSource
     ): Promise<FixupTask | null> {
-        const task = await this.typingUI.show(args, source)
+        const input = await getInput(
+            document,
+            {
+                initialRange: range,
+                initialExpandedRange: expandedRange,
+                initialModel: model,
+                initialIntent: intent,
+            },
+            source
+        )
+        if (!input) {
+            return null
+        }
+
+        const task = this.createTask(
+            document,
+            input.instruction,
+            input.userContextFiles,
+            input.range,
+            input.intent,
+            mode,
+            input.model,
+            source,
+            contextMessages
+        )
+
+        // Return focus to the editor
+        void vscode.window.showTextDocument(document)
+
         return task
     }
 
@@ -183,23 +205,21 @@ export class FixupController
         instruction: string,
         userContextFiles: ContextFile[],
         selectionRange: vscode.Range,
-        intent?: EditIntent,
-        mode: EditMode = 'edit',
+        intent: EditIntent,
+        mode: EditMode,
+        model: EditSupportedModels,
         source?: ChatEventSource,
         contextMessages?: ContextMessage[]
     ): Promise<FixupTask> {
         const fixupFile = this.files.forUri(document.uri)
-        // Support expanding the selection range for intents where it is useful
-        if (intent !== 'add') {
-            selectionRange = await getEditSmartSelection(document, selectionRange)
-        }
         const task = new FixupTask(
             fixupFile,
             instruction,
             userContextFiles,
-            intent ?? 'edit',
+            intent,
             selectionRange,
             mode,
+            model,
             source,
             contextMessages
         )
@@ -1039,36 +1059,45 @@ export class FixupController
         if (!task) {
             return
         }
-        const previousRange = task.originalRange
-        const previousInstruction = task.instruction
-        const previousUserContextFiles = task.userContextFiles
-        const document = await vscode.workspace.openTextDocument(task.fixupFile.uri)
 
+        const document = await vscode.workspace.openTextDocument(task.fixupFile.uri)
         // Prompt the user for a new instruction, and create a new fixup
-        const input = await this.typingUI.getInputFromQuickPick({
-            filePath: displayPath(task.fixupFile.uri),
-            range: previousRange,
-            initialValue: previousInstruction,
-            initialSelectedContextFiles: previousUserContextFiles,
-            source: 'code-lens',
-        })
+        const input = await getInput(
+            document,
+            {
+                initialInputValue: task.instruction,
+                initialRange: task.selectionRange,
+                initialSelectedContextFiles: task.userContextFiles,
+                initialModel: task.model,
+                initialIntent: task.intent,
+            },
+            'code-lens'
+        )
         if (!input) {
             return
         }
 
+        /**
+         * If the selected range is the same as what we provided, we actually want the original
+         * range, which is the range which will be left in the document after the task is undone.
+         *
+         * Otherwise, use the new selected range.
+         */
+        const updatedRange = input.range.isEqual(task.selectionRange) ? task.originalRange : input.range
+
         // Revert and remove the previous task
         await this.undoTask(task)
 
-        void vscode.commands.executeCommand(
-            'cody.command.edit-code',
+        void executeEdit(
             {
-                range: previousRange,
+                range: updatedRange,
                 instruction: input.instruction,
                 userContextFiles: input.userContextFiles,
                 document,
-                intent: task.intent,
+                intent: input.intent,
                 mode: task.mode,
-            } satisfies ExecuteEditArguments,
+                model: input.model,
+            },
             'code-lens'
         )
     }